@@ -1,55 +1,48 @@
 {
   "name": "collision.js",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Simple 2D JavaScript collision detection library",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
-    "build": "npm run compile",
-    "compile": "rm -rf ./lib && tsc",
-    "compile:watch": "rm -rf ./lib && tsc -w",
-    "clean": "gts clean",
-    "fix": "gts fix",
-    "lint": "gts lint",
-    "prebuild": "npm run lint && npm t",
-    "prepare": "npm run compile",
-    "pretest": "npm run compile",
-    "posttest": "npm run lint",
-    "test": "jest",
+    "fix": "npm run format:fix && npm run lint:fix",
+    "format": "prettier --check '**/*.{js,jsx,ts,tsx,json,css,scss,md}'",
+    "format:fix": "prettier --write '**/*.{js,jsx,ts,tsx,json,css,scss,md}'",
+    "lint": "eslint 'src/**/*.{js,jsx,ts,tsx}'",
+    "lint:fix": "eslint 'src/**/*.{js,jsx,ts,tsx}' --fix",
+    "prepublishOnly": "npm run format && npm run lint && npm run test",
+    "test": "node --experimental-vm-modules ./node_modules/.bin/jest",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch"
   },
+  "author": "Jesse Palmer",
+  "license": "MIT",
+  "devDependencies": {
+    "@babel/core": "^7.22.11",
+    "@babel/preset-env": "^7.22.10",
+    "babel-jest": "^29.6.4",
+    "eslint": "^8.48.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "jest": "^29.6.4",
+    "prettier": "^3.0.2"
+  },
+  "engines": {
+    "node": ">16.0.0"
+  },
+  "bugs": {
+    "url": "https://github.com/jesselpalmer/collision.js/issues"
+  },
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git://github.com/jesselpalmer/collision.js.git"
   },
   "keywords": [
-    "canvas",
-    "HTML5",
     "games",
     "collision",
     "detection"
   ],
-  "author": "Jesse Palmer",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/jesselpalmer/collision.js/issues"
-  },
-  "engines": {
-    "node": ">16.0.0"
-  },
-  "homepage": "https://github.com/jesselpalmer/collision.js",
-  "devDependencies": {
-    "@types/jest": "^29.2.4",
-    "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.47.1",
-    "@typescript-eslint/parser": "^5.47.1",
-    "eslint": "^8.30.0",
-    "gts": "^3.1.0",
-    "jest": "^29.3.1",
-    "ts-jest": "^29.0.3",
-    "ts-node": "10.9.1",
-    "typescript": "^4.0.3"
-  },
   "funding": {
     "type": "github",
     "url": "https://github.com/sponsors/jesselpalmer"
