@@ -1,4 +1,5 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
+export default {
+  transform: {
+    '^.+\\.[t|j]sx?$': 'babel-jest',
+  },
 };
